requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psutil==5.4.8
python-dotenv==0.10.1
redis==3.0.1
Werkzeug==0.14.1
